+++ crates/acts/src/cache/cache.rs
@@ -3,7 +3,7 @@ use crate::{
     Config, Result,
     data::MessageStatus,
     scheduler::{Process, Runtime, Task},
-    store::Store,
+    store::{KvStore, MemoryStore, Store},
 };
 use moka::sync::Cache as MokaCache;
 use std::{collections::HashSet, sync::Arc};
@@ -26,8 +26,9 @@ impl std::fmt::Debug for Cache {
 }
 
 impl Cache {
-    pub fn new(config: &Config) -> crate::Result<Self> {
-        let store = Arc::new(Store::create(config)?);
+    pub fn new(config: &Config, store: Option<Arc<dyn KvStore>>) -> crate::Result<Self> {
+        let store =
+            Arc::new(Store::new(store.unwrap_or_else(|| Arc::new(MemoryStore::new()))));
         Ok(Self {
             cap: config.cache_cap() as usize,
             procs: MokaCache::new(config.cache_cap() as u64),